(feat) added Modal to classroomDashboard
